refactor: split index.js into submodules
