+++ chatgpt.js
@@ -1,0 +1,25 @@
+async function createChatGPTAPI(sessionToken) {
+  // To use ESM in CommonJS, you can use a dynamic import
+  const { ChatGPTAPI } = await import("chatgpt");
+
+  const api = new ChatGPTAPI({ sessionToken });
+
+  // ensure the API is properly authenticated
+  await api.ensureAuth();
+
+  return api;
+}
+
+async function callChatGPT(api, content, retryOn503) {
+  let cnt = 0;
+  while (cnt++ <= retryOn503) {
+    try {
+      const response = await api.sendMessage(content);
+      return response;
+    } catch (err) {
+      if (!toString(err).includes("503")) throw err;
+    }
+  }
+}
+
+module.exports = { createChatGPTAPI, callChatGPT };
